Refactor client side code
  client쪽 코드에서 로그인 관련 코드를
  auth.js로 옮김
